Place Simple vue components under separate folder
