+++ src/components/simple/NumericPad.vue
@@ -19,7 +19,7 @@ export default {
   },
   methods:{
     buttonPressed(buttonId) {
-			this.$store.commit('inputHandler/handle', buttonId);
+			this.$store.commit('simple/inputHandler/handle', buttonId);
     }
   }
 }
