+++ src/components/simple/OutputControlPad.vue
@@ -17,7 +17,7 @@ export default {
   },
   methods:{
     buttonPressed(buttonId) {
-			this.$store.commit('inputHandler/handle', buttonId);
+			this.$store.commit('simple/inputHandler/handle', buttonId);
     }
   }
 }
